% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{countries_sf}
\alias{countries_sf}
\title{Countries}
\format{
\subsection{\code{countries_sf}}{

A data frame with 134,492 rows and 9 columns:
\describe{
\item{lat}{Latitude of QuadKey centroid}
\item{lon}{Longitude of QuadKey centroid}
\item{quadkey}{QuadKey number as character}
\item{country}{Country name}
\item{date_time}{Date in format \%Y-\%m-\%d \%H\%M}
\item{n_crisis}{Variable}
\item{percent_change}{Variable}
\item{day}{Day in format \%Y-\%m-\%d}
\item{time}{Hour of the day as a number between 1 and 24}
}
}
}
\usage{
countries_sf
}
\description{
Country polygons for the app example in the sf format.
}
\keyword{datasets}
