% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lakes_country}
\alias{lakes_country}
\title{Lakes from Hydrolakes DB}
\format{
\subsection{\code{lakes_country}}{

A data frame with 134,492 rows and 9 columns:
\describe{
\item{lat}{Latitude of QuadKey centroid}
\item{lon}{Longitude of QuadKey centroid}
\item{quadkey}{QuadKey number as character}
\item{country}{Country name}
\item{date_time}{Date in format \%Y-\%m-\%d \%H\%M}
\item{n_crisis}{Variable}
\item{percent_change}{Variable}
\item{day}{Day in format \%Y-\%m-\%d}
\item{time}{Hour of the day as a number between 1 and 24}
}
}
}
\usage{
lakes_country
}
\description{
Extract of the dataframe of Hydrolakes DB.
}
\keyword{datasets}
