% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Launches the app}
\usage{
run_app()
}
\value{
either run the app as a side effect or return a shiny.appobj object
}
\description{
Launches the app
}
\examples{
\dontrun{
run_app()
}
}
