% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sites_country}
\alias{sites_country}
\title{sites country}
\format{
\subsection{\code{sites_country}}{

A data frame with 360 rows and 2 columns:
\describe{
\item{quadkey}{QuadKey number as character}
\item{variable}{Numeric variable}
}
}
}
\usage{
sites_country
}
\description{
A data frame simulating a variable asociated with QuadKey numbers.
}
\keyword{datasets}
